Refactor user authentication and registration flow
- Remove AuthenticateUser and UserController
- Move authentication logic to UserAction
- Add logout route and controller method
- Create RegisterController for registration and authentication
- Update routes and AuthController accordingly
- Add .DS_Store files to Actions and src directories
