+++ modules/User/routes/routes.php
@@ -10,8 +10,13 @@ Route::middleware(["guest", "throttle:10,1"])->group(function () {
     Route::post("/register", [AuthController::class, "store"])->name(
         "register"
     );
+
     Route::post("/sanctum/token", [
         AuthController::class,
         "authenticate",
     ])->name("token");
+});
+
+Route::middleware(["auth:sanctum", "throttle:10,1"])->group(function () {
+    Route::post("/logout", [AuthController::class, "logout"])->name("logout");
 });
