+++ modules/User/src/Actions/UserAction.php
@@ -1,8 +1,10 @@
 <?php
 namespace Modules\User\Actions;
 
+use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Str;
 use Modules\User\Events\UserCreated;
+use Modules\User\Models\User;
 
 class UserAction
 {
@@ -20,5 +22,47 @@ class UserAction
         $attributes["uuid"] = (string) Str::uuid()->toString();
 
         event(new UserCreated($attributes));
+    }
+
+    /**
+     * Handle the authentication of a user.
+     *
+     * @param  array  $attributes
+     * @return string
+     *
+     * @throws \Illuminate\Validation\ValidationException
+     */
+    public function createToken(array $attributes)
+    {
+        $user = $this->getUserByEmail($attributes["email"]);
+
+        if ($user && Hash::check($attributes["password"], $user->password)) {
+            //fire an event: UserLoggedIn
+
+            return $user->createToken("auth_token")->plainTextToken;
+        }
+        return false;
+    }
+
+    public function clearToken($request)
+    {
+        //fire logged out event
+
+        $request->auth();
+
+        $request->session()->invalidate();
+
+        $request->session()->regenerateToken();
+    }
+
+    /**
+     * Get the user by email.
+     *
+     * @param  string  $email
+     * @return User|null
+     */
+    private function getUserByEmail(string $email): ?User
+    {
+        return User::where("email", $email)->first();
     }
 }
