--- modules/User/src/Http/Controllers/UserController.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Modules\User\Http\Controllers;
-
-use Modules\User\Http\Requests\CreateUserRequest;
-use Modules\User\Models\User;
-
-class UserController {}
