+++ modules/User/src/Models/User.php
@@ -2,6 +2,7 @@
 
 namespace Modules\User\Models;
 
+// use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Database\Eloquent\Concerns\HasUuids;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Notifications\Notifiable;
